доделал внешний вид гл.страницы

+++ MappingProfile.cs
@@ -14,7 +14,7 @@ namespace SocialNetwork
         {
             CreateMap<RegisterViewModel, User>()
                 .ForMember(dest => dest.UserName, opt => opt.MapFrom(src => src.FirstName))
-                 .ForMember(dest => dest.BirthDate, opt => opt.MapFrom(src => new DateTime(src.Year, src.Month, src.Date)))
+                 .ForMember(dest => dest.BirthDate, opt => opt.MapFrom(src => new DateTime(src.Year, (int)src.Month, src.Date)))
                  .ForMember(dest => dest.Email, opt => opt.MapFrom(src => src.EmailReg));
